Fix and validate MCP (Master Control Program) system - resolve circular imports and modernize datetime usage (#24)
* Initial plan

* Fix circular imports and test MCP functionality - all endpoints working


* Complete MCP system validation - add documentation and test scripts


---------
