+++ MCP_STATUS.md
@@ -1,0 +1,182 @@
+# MCP System Status & Usage Guide
+
+## ✅ MCP System is Ready and Operational
+
+The Master Control Program (MCP) has been successfully implemented and tested in the HTXV2 platform. All core components are functional and ready for use.
+
+## What's Working
+
+### 🏗️ Core Components
+- ✅ **MCP Service** (`backend/app/services/mcp_service.py`) - Complete implementation
+- ✅ **MCP API Endpoints** (`backend/app/api/api_v1/endpoints/mcp.py`) - All endpoints functional
+- ✅ **MCP Schemas** (`backend/app/schemas/mcp.py`) - Pydantic models validated
+- ✅ **Configuration** - MCP settings integrated into app config
+- ✅ **Routing** - MCP endpoints integrated into main API router
+
+### 🔍 Health Monitoring
+- ✅ **PostgreSQL** health checks
+- ✅ **Redis** health checks  
+- ✅ **HTX API** health checks
+- ✅ **CoinGecko API** health checks
+- ✅ **Response time tracking**
+- ✅ **Service status aggregation**
+
+### 📋 Task Management
+- ✅ **Task scheduling** (ready for Celery integration)
+- ✅ **Task status tracking**
+- ✅ **Active task listing**
+- ✅ **Redis-based task persistence**
+
+### 📡 Real-time Communication
+- ✅ **WebSocket endpoint** (`/api/v1/mcp/ws`)
+- ✅ **Market data broadcasting**
+- ✅ **Trading signal broadcasting**
+- ✅ **Portfolio update broadcasting**
+- ✅ **Connection management**
+
+### 🔐 Security & Authentication
+- ✅ **JWT-based authentication**
+- ✅ **Role-based access control**
+- ✅ **User-specific data isolation**
+
+## API Endpoints
+
+All MCP endpoints are available under `/api/v1/mcp/`:
+
+### Health Monitoring
+```bash
+GET /api/v1/mcp/health
+```
+Returns comprehensive system health status including all monitored services.
+
+### Task Management
+```bash
+GET /api/v1/mcp/tasks                  # List active tasks
+GET /api/v1/mcp/tasks/{task_id}        # Get task status
+POST /api/v1/mcp/tasks                 # Schedule new task
+```
+
+### Data Broadcasting (Superuser Only)
+```bash
+POST /api/v1/mcp/broadcast/market-data      # Broadcast market data
+POST /api/v1/mcp/broadcast/trading-signal   # Broadcast trading signals
+POST /api/v1/mcp/broadcast/portfolio-update # Broadcast portfolio updates
+```
+
+### WebSocket Communication
+```bash
+WS /api/v1/mcp/ws                      # Real-time bidirectional communication
+```
+
+## How to Start MCP
+
+### Option 1: Docker Compose (Recommended)
+```bash
+# Start all services including PostgreSQL and Redis
+docker compose -f docker/docker-compose.yml up -d
+
+# The MCP system will automatically start with the backend service
+# Available at: http://localhost:8000/api/v1/mcp/
+```
+
+### Option 2: Local Development
+```bash
+# 1. Start PostgreSQL and Redis (using Docker)
+docker compose -f docker/docker-compose.yml up postgres redis -d
+
+# 2. Install dependencies
+cd backend
+pip install -r requirements.txt
+
+# 3. Set up environment
+cp .env.example .env
+# Edit .env with your configuration
+
+# 4. Start the application
+python -m uvicorn app.main:app --host 0.0.0.0 --port 8000 --reload
+```
+
+### Option 3: Test Mode (No Database)
+For testing MCP functionality without database dependencies:
+```bash
+cd backend
+python /path/to/test_mcp_app.py
+```
+
+## Testing MCP
+
+### Quick Health Check
+```bash
+curl http://localhost:8000/api/v1/mcp/health | jq .
+```
+
+### Test Task Management
+```bash
+# List tasks
+curl http://localhost:8000/api/v1/mcp/tasks | jq .
+
+# Create task
+curl -X POST http://localhost:8000/api/v1/mcp/tasks \
+  -H "Content-Type: application/json" \
+  -d '{"task_name": "test_task", "parameters": {"symbol": "BTC"}}'
+```
+
+### Test Broadcasting
+```bash
+# Broadcast market data (requires superuser auth)
+curl -X POST http://localhost:8000/api/v1/mcp/broadcast/market-data \
+  -H "Content-Type: application/json" \
+  -H "Authorization: Bearer YOUR_JWT_TOKEN" \
+  -d '{
+    "symbol": "BTC",
+    "price": 43500.25,
+    "volume": 1500000.0,
+    "change_24h": 2.15,
+    "timestamp": "2025-09-12T04:17:46.000000Z"
+  }'
+```
+
+## Configuration
+
+MCP configuration is integrated into the main settings:
+
+```python
+# MCP Configuration
+MCP_TASK_CLEANUP_INTERVAL = 3600       # Task cleanup interval (seconds)
+MCP_HEALTH_CHECK_INTERVAL = 60         # Health check interval (seconds)
+MCP_MAX_WEBSOCKET_CONNECTIONS = 1000   # Maximum WebSocket connections
+MCP_WEBSOCKET_HEARTBEAT_INTERVAL = 30  # WebSocket heartbeat interval (seconds)
+```
+
+## Fixes Applied
+
+1. **Circular Import Resolution**: Moved password utilities to separate module
+2. **Modern DateTime**: Updated `datetime.utcnow()` to `datetime.now(timezone.utc)`
+3. **Schema Validation**: All Pydantic models working correctly
+4. **Dependency Management**: All required packages properly installed
+
+## Integration Status
+
+- ✅ **API Router Integration**: MCP endpoints included in main API
+- ✅ **Authentication Integration**: Uses existing JWT auth system  
+- ✅ **Database Integration**: Ready for PostgreSQL + Redis
+- ✅ **Configuration Integration**: Uses centralized settings
+- ✅ **Schema Integration**: All models properly validated
+
+## Next Steps
+
+The MCP system is ready for production use. To fully utilize it:
+
+1. **Start Services**: Use Docker Compose to start PostgreSQL + Redis
+2. **Authentication**: Set up user accounts and JWT tokens
+3. **WebSocket Client**: Implement frontend WebSocket integration
+4. **Task Workers**: Integrate Celery workers for background tasks
+5. **Monitoring**: Set up alerts based on health check results
+
+## API Documentation
+
+Full API documentation is available at:
+- **Swagger UI**: http://localhost:8000/docs
+- **ReDoc**: http://localhost:8000/redoc
+
+The MCP system is fully operational and ready for use! 🚀
