+++ backend/app/services/user_service.py
@@ -5,7 +5,7 @@ from sqlalchemy.sql import func
 
 from app.models.user import User
 from app.schemas.user import UserCreate, UserUpdate
-from app.core.security import get_password_hash, verify_password
+from app.core.password import get_password_hash, verify_password
 
 
 class UserService:
